Add registered and library kernels
